requirements.backend.txt
fastapi==0.109.1
uvicorn==0.27.0
pydantic==2.6.1
requests==2.31.0
llama-cpp-python==0.2.90
langchain==0.1.6
langchain-community==0.0.19
chromadb==0.4.22
sentence-transformers==2.2.2
huggingface_hub==0.20.3
pymupdf==1.23.7
